requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.31.0
accelerate>=0.21.0
xformers>=0.0.20
bitsandbytes>=0.41.0
Pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.15.0
safetensors>=0.3.0
peft>=0.5.0
datasets>=2.14.0